Reorganize account manager creation
