+++ src/helper.rs
@@ -3,11 +3,10 @@
 
 use std::path::Path;
 
-use clap::Parser;
 use dialoguer::{console::Term, theme::ColorfulTheme, Password, Select};
 use iota_wallet::account::AccountHandle;
 
-use crate::{error::Error, AccountManagerCli};
+use crate::error::Error;
 
 pub fn get_password(path: &Path) -> Result<String, Error> {
     let mut prompt = Password::new();
@@ -35,21 +34,4 @@ pub async fn pick_account(accounts: Vec<AccountHandle>) -> Option<usize> {
         .default(0)
         .interact_on_opt(&Term::stderr())
         .unwrap_or_default()
-}
-
-pub fn help_command() {
-    if let Err(r) = AccountManagerCli::try_parse() {
-        // If only one argument from the user is provided, try to use it as identifier.
-        let mut iter = std::env::args();
-
-        // The first element is the path of the executable.
-        iter.next();
-        if let Some(input) = iter.next() {
-            if input == "help" {
-                // this prints the help output
-                r.print().expect("Error writing Error");
-                std::process::exit(0);
-            }
-        }
-    }
 }
